python-ml-service/requirements.txt
kafka-python>=2.0.2
numpy>=1.24.0
scikit-learn>=1.2.0
pandas>=2.0.0
joblib>=1.2.0
flask>=2.3.0
flask-cors>=4.0.0
python-dateutil>=2.8.2
